pyproject.toml
[tool.poetry]
name = "pokemon-capstone-backend"
version = "0.1.0"
description = ""
authors = ["Paul <paulieracisz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
djangorestframework = "^3.12.1"
django-cors-headers = "^3.5.0"
djangorestframework-jwt = "^1.11.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
